+++ addons/asset_placer/ui/settings/settings_panel.gd
@@ -22,6 +22,8 @@ var _presenter: SettingsPresenter = SettingsPresenter.new()
 @onready var keybinding_option_axis_y = %KeybindingOptionAxisY
 @onready var keybinding_option_axis_z = %KeybindingOptionAxisZ
 @onready var keybinding_option_plane_mode = $Panel/MarginContainer/ScrollContainer/VBoxContainer/HBoxContainer/KeyBindings/KeyBindingsOptions/KeybindingOptionPlaneMode
+@onready var update_channel_option_button: OptionButton = %UpdateChannelOptionButton
+@onready var update_channel_info_button: Button = %UpdateChannelInfoButton
 
 
 func _ready():
@@ -29,6 +31,7 @@ func _ready():
 	trasform_step_spin_box.value_changed.connect(_presenter.set_default_transform_step)
 	rotation_step_spin_box.value_changed.connect(_presenter.set_rotation_step)
 	_presenter.show_settings.connect(_show_settings)
+	update_channel_option_button.item_selected.connect(_presenter.set_update_channel)
 	keybinding_option_rotate.keybind_changed.connect(func(key):
 		_presenter.set_binding(AssetPlacerSettings.Bindings.Rotate, key)
 	)
@@ -69,6 +72,11 @@ func _ready():
 	reset_button.pressed.connect(_presenter.reset_to_defaults)
 	material_clear_button.pressed.connect(_presenter.clear_preivew_material)
 	material_picker_button.pressed.connect(_show_preview_material_picker)
+	
+	update_channel_info_button.pressed.connect(func():
+		OS.shell_open("https://levinzonr.github.io/godot-asset-placer/development-lifecycle/")
+	)
+	
 	plane_material_picker_button.pressed.connect(func():
 		EditorInterface.popup_quick_open(_presenter.set_plane_material, ["BaseMaterial3D"])
 	)
@@ -90,7 +98,7 @@ func _show_settings(setting: AssetPlacerSettings):
 	keybinding_option_axis_y.set_keybind(setting.bindings[AssetPlacerSettings.Bindings.ToggleAxisY])
 	keybinding_option_axis_z.set_keybind(setting.bindings[AssetPlacerSettings.Bindings.ToggleAxisZ])
 	keybinding_option_plane_mode.set_keybind(setting.bindings[AssetPlacerSettings.Bindings.TogglePlaneMode])
-
+	update_channel_option_button.select(setting.update_channel)
 	plane_material_picker_button.text = setting.plane_material_resource.get_file()
 	if setting.preview_material_resource.is_empty():
 		material_picker_button.text = "No Preview Material"